refactor: 대시보드 페이지 데이터 타입 정의
