+++ src/service/dashboard/types.d.ts
@@ -1,16 +1,14 @@
-import { EmotionKey } from '..';
+import type { EmotionKey } from '@models/index';
 
-export type DailyAmountType = {
+type DailyAmountType = {
   date: string;
   amount: number;
 };
-
-export type EmotionAmountTotalType = {
+type EmotionAmountTotalType = {
   emotion: EmotionKey;
   amount: number;
 };
-
-export type DashboardPageDataType = {
+type DashboardDataType = {
   dailyAmount: DailyAmountType[];
   emotionAmountTotal: EmotionAmountTotalType[];
   satisfactionAverage: number;
